requirements.txt
requests
PySocks
rich
tabulate
geoip2
speedtest-cli
pyyaml
